+++ src/main/java/pokeraidbot/commands/WhatsNewCommand.java
@@ -22,6 +22,12 @@ public class WhatsNewCommand extends ConfigAwareCommand {
         final String message;
         if (config.getLocale().equals(LocaleService.SWEDISH)) {
             message = "**Nytt i 1.6.0-" + BotServerMain.version + ":**\n\n" +
+                    "- Groudon ny default tier5-raidboss 2017-12-15\n" +
+                    "- Ho-Oh är inte längre default tier5-raidboss 2017-12-14\n" +
+                    "- Möjligt för bot admin att ändra nuvarande tier5 boss i runtime\n" +
+                    "- Mindre korrigering av i18n texter\n" +
+                    "- Ny import av kartor (ändringar)\n" +
+                    "- Fix: Om en admin misslyckats med !raid install, kommer inte längre loggen spammas\n" +
                     "- Fix: !raid track när en server har botintegration aktiverad ska nu funka som det ska\n" +
                     "- !raid group titel har nu raidstart och -slut\n" +
                     "- Fix: Eventlyssnare för grupper rensas undan bättre\n" +
@@ -37,6 +43,12 @@ public class WhatsNewCommand extends ConfigAwareCommand {
                     "- Fix: EX raids kommer inte längre skapas via botintegration";
         } else {
             message = "**New in 1.6.0-" + BotServerMain.version + ":**\n\n" +
+                    "- Groudon new default tier5-raidboss 2017-12-15\n" +
+                    "- Ho-Oh no longer default tier 5 boss after 2017-12-14\n" +
+                    "- Possible to change current tier5 pokemon for bot admin on the fly\n" +
+                    "- Minor fixes to i18n texts\n" +
+                    "- Fresh maps imported due to changes\n" +
+                    "- Fix: If a server admin failed with !raid install, they will no longer spam the log needlessly\n" +
                     "- Fix: !raid track when a server has bot integration setting active should now work properly\n" +
                     "- Add raid duration to group message header\n" +
                     "- Fix: Better cleanup of event listeners\n" +